requirements.txt
torch
torchvision
transformers>=4.45.0
pillow>=9.0.0
accelerate>=0.20.0
